+++ client/src/actions/accounts.js
@@ -14,5 +14,3 @@ export const balanceOfCln = (accountAddress) => action(BALANCE_OF_CLN.REQUEST, {
 
 export const fetchBalances = (tokens, accountAddress) => action(FETCH_BALANCES.REQUEST, {accountAddress, tokens})
 export const fetchTokensWithBalances = (accountAddress) => action(FETCH_TOKENS_WITH_BALANCES.REQUEST, {accountAddress})
-
-export const addUserInformation = (user) => action(ADD_USER_INFORMATION.REQUEST, {user})
